Uses pysc2 api to make units move
All player units are one by one selected and then moved to a new location automatically. This location is dependent on the enemy's location, as the unit is given a code 0 for backward(away from enemy), 1 for forward (toward enemy), and 2 for attack.
Each units location is found via KMeans Clustering, to get the center pixel of the unit. These coordinates are stored as variables in the agent's class, and are only updated after all units on the screen have been moved.
Enemy location is found through rounding at the moment, as we are assuming only one enemy.